+++ src/main/java/uk/gov/pay/connector/app/ConnectorApp.java
@@ -23,9 +23,12 @@ import io.dropwizard.setup.Bootstrap;
 import io.dropwizard.setup.Environment;
 import org.glassfish.jersey.server.filter.RolesAllowedDynamicFeature;
 import uk.gov.pay.commons.utils.xray.Xray;
+import uk.gov.pay.connector.cardtype.resource.CardTypesResource;
 import uk.gov.pay.connector.charge.resource.ChargesApiResource;
 import uk.gov.pay.connector.charge.resource.ChargesFrontendResource;
+import uk.gov.pay.connector.chargeevents.resource.ChargeEventsResource;
 import uk.gov.pay.connector.command.RenderStateTransitionGraphCommand;
+import uk.gov.pay.connector.report.resource.TransactionsSummaryResource;
 import uk.gov.pay.connector.exception.ConstraintViolationExceptionMapper;
 import uk.gov.pay.connector.exception.ValidationExceptionMapper;
 import uk.gov.pay.connector.filters.LoggingFilter;
@@ -36,22 +39,19 @@ import uk.gov.pay.connector.gatewayaccount.resource.GatewayAccountResource;
 import uk.gov.pay.connector.healthcheck.CardExecutorServiceHealthCheck;
 import uk.gov.pay.connector.healthcheck.DatabaseHealthCheck;
 import uk.gov.pay.connector.healthcheck.Ping;
-import uk.gov.pay.connector.resources.CardResource;
-import uk.gov.pay.connector.resources.CardTypesResource;
-import uk.gov.pay.connector.chargeevents.resource.ChargeEventsResource;
+import uk.gov.pay.connector.healthcheck.resource.HealthCheckResource;
 import uk.gov.pay.connector.refund.resource.ChargeRefundsResource;
-import uk.gov.pay.connector.usernotification.resource.EmailNotificationResource;
-import uk.gov.pay.connector.resources.HealthCheckResource;
-import uk.gov.pay.connector.webhooks.resource.NotificationResource;
-import uk.gov.pay.connector.resources.PerformanceReportResource;
 import uk.gov.pay.connector.refund.resource.SearchRefundsResource;
-import uk.gov.pay.connector.resources.SecurityTokensResource;
-import uk.gov.pay.connector.resources.TransactionsSummaryResource;
+import uk.gov.pay.connector.report.resource.PerformanceReportResource;
+import uk.gov.pay.connector.resources.CardResource;
 import uk.gov.pay.connector.service.CaptureProcessScheduler;
 import uk.gov.pay.connector.service.CardCaptureProcess;
+import uk.gov.pay.connector.token.resource.SecurityTokensResource;
+import uk.gov.pay.connector.usernotification.resource.EmailNotificationResource;
 import uk.gov.pay.connector.util.DependentResourceWaitCommand;
 import uk.gov.pay.connector.util.TrustingSSLSocketFactory;
 import uk.gov.pay.connector.util.XrayUtils;
+import uk.gov.pay.connector.webhooks.resource.NotificationResource;
 
 import javax.net.ssl.HttpsURLConnection;
 import javax.net.ssl.SSLSocketFactory;
