+++ src/main/java/uk/gov/pay/connector/cardtype/model/domain/UuidAbstractEntity.java
@@ -1,8 +1,13 @@
-package uk.gov.pay.connector.model.domain;
+package uk.gov.pay.connector.cardtype.model.domain;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 
-import javax.persistence.*;
+import javax.persistence.Column;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+import javax.persistence.MappedSuperclass;
+import javax.persistence.Version;
 import java.io.Serializable;
 import java.util.UUID;
