+++ src/main/java/uk/gov/pay/connector/gateway/model/PayersCardType.java
@@ -1,6 +1,6 @@
 package uk.gov.pay.connector.gateway.model;
 
-import uk.gov.pay.connector.model.domain.CardTypeEntity;
+import uk.gov.pay.connector.cardtype.model.domain.CardTypeEntity;
 
 /**
  * The enum type that should be used to map from and to JSON when
