+++ src/main/java/uk/gov/pay/connector/gatewayaccount/model/PatchRequest.java
@@ -1,4 +1,4 @@
-package uk.gov.pay.connector.model;
+package uk.gov.pay.connector.gatewayaccount.model;
 
 import com.fasterxml.jackson.core.type.TypeReference;
 import com.fasterxml.jackson.databind.JsonNode;
