+++ src/main/java/uk/gov/pay/connector/healthcheck/resource/HealthCheckResource.java
@@ -1,4 +1,4 @@
-package uk.gov.pay.connector.resources;
+package uk.gov.pay.connector.healthcheck.resource;
 
 import com.codahale.metrics.health.HealthCheck;
 import com.fasterxml.jackson.core.JsonProcessingException;
