+++ src/main/java/uk/gov/pay/connector/report/resource/TransactionsSummaryResource.java
@@ -1,14 +1,14 @@
-package uk.gov.pay.connector.resources;
+package uk.gov.pay.connector.report.resource;
 
 import org.apache.commons.lang3.tuple.Pair;
 import uk.gov.pay.connector.charge.dao.ChargeDao;
 import uk.gov.pay.connector.charge.model.domain.ChargeEntity;
-import uk.gov.pay.connector.refund.dao.RefundDao;
+import uk.gov.pay.connector.charge.model.domain.ChargeStatus;
+import uk.gov.pay.connector.report.model.TransactionsSummaryResponse;
 import uk.gov.pay.connector.gatewayaccount.dao.GatewayAccountDao;
-import uk.gov.pay.connector.model.TransactionsSummaryResponse;
 import uk.gov.pay.connector.model.api.ExternalChargeState;
 import uk.gov.pay.connector.model.api.ExternalRefundStatus;
-import uk.gov.pay.connector.charge.model.domain.ChargeStatus;
+import uk.gov.pay.connector.refund.dao.RefundDao;
 import uk.gov.pay.connector.refund.model.domain.RefundEntity;
 import uk.gov.pay.connector.refund.model.domain.RefundStatus;
 import uk.gov.pay.connector.util.ResponseUtil;
