+++ src/main/java/uk/gov/pay/connector/token/dao/TokenDao.java
@@ -1,9 +1,9 @@
-package uk.gov.pay.connector.dao;
+package uk.gov.pay.connector.token.dao;
 
 import com.google.inject.Provider;
 import com.google.inject.persist.Transactional;
 import uk.gov.pay.connector.common.dao.JpaDao;
-import uk.gov.pay.connector.model.domain.TokenEntity;
+import uk.gov.pay.connector.token.model.domain.TokenEntity;
 
 import javax.inject.Inject;
 import javax.persistence.EntityManager;
