+++ src/test/java/uk/gov/pay/connector/it/dao/GuicedTestEnvironment.java
@@ -8,8 +8,8 @@ import com.google.inject.persist.PersistService;
 import com.google.inject.persist.jpa.JpaPersistModule;
 import uk.gov.pay.connector.charge.dao.ChargeDao;
 import uk.gov.pay.connector.chargeevents.dao.ChargeEventDao;
-import uk.gov.pay.connector.dao.TokenDao;
 import uk.gov.pay.connector.gatewayaccount.dao.GatewayAccountDao;
+import uk.gov.pay.connector.token.dao.TokenDao;
 
 public class GuicedTestEnvironment {
