+++ src/test/java/uk/gov/pay/connector/it/dao/SupportedTypeDaoJpaITest.java
@@ -2,15 +2,16 @@ package uk.gov.pay.connector.it.dao;
 
 import org.junit.Before;
 import org.junit.Test;
-import uk.gov.pay.connector.dao.CardTypeDao;
-import uk.gov.pay.connector.model.domain.CardTypeEntity;
+import uk.gov.pay.connector.cardtype.dao.CardTypeDao;
+import uk.gov.pay.connector.cardtype.model.domain.CardTypeEntity;
 
 import java.util.List;
 
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.core.Is.is;
 import static org.junit.Assert.assertNotNull;
-import static uk.gov.pay.connector.model.domain.CardTypeEntity.SupportedType.*;
+import static uk.gov.pay.connector.cardtype.model.domain.CardTypeEntity.SupportedType.CREDIT;
+import static uk.gov.pay.connector.cardtype.model.domain.CardTypeEntity.SupportedType.DEBIT;
 
 public class SupportedTypeDaoJpaITest extends DaoITestBase {
