+++ src/test/java/uk/gov/pay/connector/it/dao/TokenDaoJpaITest.java
@@ -5,8 +5,8 @@ import org.junit.Rule;
 import org.junit.Test;
 import org.junit.rules.ExpectedException;
 import uk.gov.pay.connector.charge.model.domain.ChargeEntity;
-import uk.gov.pay.connector.dao.TokenDao;
-import uk.gov.pay.connector.model.domain.TokenEntity;
+import uk.gov.pay.connector.token.dao.TokenDao;
+import uk.gov.pay.connector.token.model.domain.TokenEntity;
 
 import java.util.Optional;
